+++ SofaKernel/modules/SofaBaseTopology/SparseGridTopology.h
@@ -357,7 +357,8 @@ public :
         return sofa::component::topology::MeshTopology::getNbPoints();
     }
 
-    virtual int getNbHexahedra() override { return (int)this->getHexahedra().size();}
+    /// TODO 2018-07-23 epernod: check why this method is override to return the same result as parent class.
+    virtual size_t getNbHexahedra() override { return this->getHexahedra().size();}
 };
 
 } // namespace topology
